Be less strict when extending past the end of the ROM
f894ee9 allows us to append data to the end of the ROM by creating a
patch starting exactly at the end of the input. However, the Super
Metroid Practice Hack (https://smpractice.speedga.me/) places content
well past the end of the input ROM (with a sizable gap in between).

Instead of only allowing patches to strictly append to the ROM, allow
writing anywhere past the end of the ROM (zeroing any unused space).